# Conflicts:
#	Ghostbait/Ghostbait/Source Files/InputManager.cpp
#	Ghostbait/Ghostbait/Source Files/RigidBody.cpp
#	Ghostbait/Ghostbait/Source Files/VRManager.cpp

+++ Ghostbait/Ghostbait/Source Files/ObjectManager.cpp
@@ -5,7 +5,7 @@ ObjectManager::ObjectManager(MemoryManager* _memMan): memMan(_memMan)
 {
 	//TODO: Fix this into Initialize
 	poolListCount = 80;
-	poolList = (_Pool_Base* )memMan->RequestMemory(poolListCount, sizeof(Pool<size_t>));
+	poolList = (_Pool_Base* )memMan->RequestMemory((int)poolListCount, sizeof(Pool<size_t>));
 	//objectpool.reserve(prefabCount);
 }
